modules/life-management/patterns/backend/requirements.txt
# Core Flask backend
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# PDF generation
reportlab==4.0.7
pillow==10.1.0

# Data processing (minimal set)
numpy==1.24.4
pandas>=2.2.0
matplotlib==3.8.2

# API clients (keep for future flexibility)
openai==1.3.7
